Fix Category delete

+++ src/Service/CategoryService.php
@@ -3,14 +3,21 @@
 namespace App\Service;
 
 use App\Entity\Category;
+use App\Entity\Limit;
 use App\Repository\CategoryRepository;
+use App\Repository\LimitRepository;
+use App\Repository\MoneyOperationRepository;
 
 class CategoryService
 {
     private CategoryRepository $categoryRepository;
-    public function __construct(CategoryRepository $categoryRepository)
+    private MoneyOperationRepository $moneyOperationRepository;
+    private LimitRepository $limitRepository;
+    public function __construct(LimitRepository $limitRepository, CategoryRepository $categoryRepository, MoneyOperationRepository $moneyOperationRepository)
     {
         $this->categoryRepository = $categoryRepository;
+        $this->moneyOperationRepository = $moneyOperationRepository;
+        $this->limitRepository = $limitRepository;
     }
 
     public function findAllByTypeAndUserId(bool $type, int $userId)
@@ -30,6 +37,13 @@ class CategoryService
 
     public function delete(Category $category)
     {
-        $this->categoryRepository->delete($category);
+        $operations = $this->moneyOperationRepository->findAllByCategory($category);
+        $limits = $this->limitRepository->findAllByCategory($category);
+        if (count($operations) === 0 && count($limits) === 0) {
+            $this->categoryRepository->delete($category);
+            return true;
+        } else {
+            return false;
+        }
     }
 }
